requirements.txt
# Core Django
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1

# Database
mysqlclient==2.2.1

# Authentication
django-cors-headers==4.3.1

# Development
debugpy==1.8.0
django-debug-toolbar==4.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Telegram Integration
python-telegram-bot==20.7

# Code Quality
black==24.3.0
flake8==7.0.0
isort==5.13.2

# Documentation
Pillow==10.2.0
